[MOVE/#13] 패키지 생성
