+++ app/src/main/AndroidManifest.xml
@@ -29,7 +29,7 @@
             android:exported="false"
             android:windowSoftInputMode="adjustResize" />
         <activity
-            android:name=".presentation.BnvActivity"
+            android:name=".presentation.main.BnvActivity"
             android:exported="false" />
     </application>
