+++ app/src/main/java/org/sopt/dosopttemplate/presentation/main/home/HomeFragment.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate.presentation
+package org.sopt.dosopttemplate.presentation.main.home
 
 import android.os.Bundle
 import android.view.LayoutInflater
