+++ app/src/main/java/org/sopt/dosopttemplate/presentation/main/mypage/MypageFragment.kt
@@ -1,4 +1,4 @@
-package org.sopt.dosopttemplate.presentation
+package org.sopt.dosopttemplate.presentation.main.mypage
 
 import android.content.Intent
 import android.content.Intent.FLAG_ACTIVITY_CLEAR_TASK
